extract emblanavbutton component
